requirements.txt
PyYAML==3.11
SQLAlchemy==0.9.8
eventlet==0.15.2
greenlet==0.4.5
wsgiref==0.1.2